+++ algorithms/insertion_sort.h
@@ -1,1 +1,3 @@
-int *insertion_sort(int *arr, signed int size);
+#include <stdlib.h>
+
+int *insertion_sort(int *arr, size_t size);
